+++ Leomard/Views/Components/PostUIView.swift
@@ -40,6 +40,9 @@ struct PostUIView: View {
     
     @State var showFailedToFeatureAlert: Bool = false
     
+    @State var startRemovePost: Bool = false
+    @State var removalReason: String = ""
+    
     @Environment(\.openURL) var openURL
     
     var body: some View {
@@ -120,6 +123,22 @@ struct PostUIView: View {
             .contextMenu {
                 PostContextMenu(contentView: contentView, postView: self.postView, sender: self)
             }
+            .alert("Remove Post (Mod)", isPresented: $startRemovePost, actions: {
+                TextField("Optional", text: $removalReason)
+                Button("Remove", role: .destructive) {
+                    self.postService.remove(post: postView.post, reason: removalReason, removed: true) { result in
+                        switch result {
+                        case .success(let postResponse):
+                            self.postView = postResponse.postView
+                        case .failure(let error):
+                            print(error)
+                        }
+                    }
+                }
+                Button("Cancel", role: .cancel) {}
+            }, message: {
+                Text("State the reason of removal:")
+            })
         }
     }
     
@@ -182,6 +201,10 @@ struct PostUIView: View {
             if postView.post.featuredCommunity {
                 Image(systemName: "pin.fill")
                     .foregroundColor(.red)
+            }
+            if postView.post.locked {
+                Image(systemName: "lock.fill")
+                    .foregroundColor(.green)
             }
             Text(postView.post.name)
                 .bold()
@@ -575,5 +598,20 @@ struct PostUIView: View {
     
     func crossPost() {
         contentView.openCrossPost(post: postView)
+    }
+    
+    func startPostRemoval() {
+        startRemovePost = true
+    }
+    
+    func lock() {
+        postService.lock(post: postView.post, locked: !postView.post.locked) { result in
+            switch result {
+            case .success(let postResponse):
+                self.postView = postResponse.postView
+            case .failure(let error):
+                print(error)
+            }
+        }
     }
 }