+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 
-## TBA (TBA)
+## 0.4 (TBA)
+
+## Added
+
+- Mod Tools:
+    - Remove posts
+    - Lock posts
+    - Distinguish comment
+    - Remove comments
+    - Remove community
+    - Edit community
+- Post-locked indicator
+- Distinguished comment indicator
 
 ### Bug Fixes